# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/reviews/me" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "getMyReviews",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 50,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedModelReviewDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/reviews" : {
      "get" : {
        "tags" : [ "admin-review-controller" ],
        "operationId" : "getReviewsForEntity",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 50,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedModelReviewDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/reviews/by-user/{userId}" : {
      "get" : {
        "tags" : [ "admin-review-controller" ],
        "operationId" : "getReviewsByUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 50,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedModelReviewDetailDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ReviewDetailDto" : {
        "required" : [ "authorRole", "createdAt", "entityId", "entityType", "id", "orderId", "rating", "userId", "visibility" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entityType" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          },
          "entityId" : {
            "type" : "string"
          },
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string"
          },
          "authorRole" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "RESTAURANT", "DELIVERY" ]
          },
          "visibility" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "PRIVATE" ]
          },
          "authorDisplayName" : {
            "type" : "string"
          },
          "rating" : {
            "type" : "integer",
            "format" : "int32"
          },
          "comment" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PageMetadata" : {
        "required" : [ "number", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "ApiResponsePagedModelReviewDetailDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedModelReviewDetailDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PagedModelReviewDetailDto" : {
        "required" : [ "content" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewDetailDto"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        }
      },
...
    }
  }
}